+++ tests/__init__.py
@@ -1,0 +1,7 @@
+"""All the tests for the jira package.
+
+Refer to conftest.py for shared helper methods.
+
+resources/test_* : For tests related to resources
+test_* : For other tests of the non-resource elements of the jira package.
+"""
