+++ tests/resources/__init__.py
@@ -1,0 +1,5 @@
+"""Tests grouped by Resource type
+
+The resources/ folder contains tests grouped per
+jira.resource.Resource with files for each of its subclasses.
+"""
